Python Gmail Program
This program was made using the Tkinter and smtplib/MIME libraries in Python. The program connect to the Gmail server, and asks the user for their user credentials. It then allows for an email subject, text and attachment to be added. The program then sends the email to the intended user instantaneously.